[python] Organize LeRobot importer as a package
